moved files index and types from src
