+++ index.ts
@@ -1,0 +1,6 @@
+import { validateTx } from "./src/utils/validateTx"
+import { onPayment } from "./src/utils/onPayment"
+import { generatePaymentLink } from "./src/utils/generatePaymentLink";
+import { generatePaymentQR } from "./src/utils/generatePaymentQR";
+
+export {validateTx, onPayment, generatePaymentLink, generatePaymentQR};
