--- src/index.ts
@@ -1,6 +1,0 @@
-import { validateTx } from "./utils/validateTx"
-import { onPayment } from "./utils/onPayment"
-import { generatePaymentLink } from "./utils/generatePaymentLink";
-import { generatePaymentQR } from "./utils/generatePaymentQR";
-
-export {validateTx, onPayment, generatePaymentLink, generatePaymentQR};
